JITM: Move the JITM::register call to the Jetpack class
Package classes must not be used before all of the plugins have loaded.
Move the instantiation of the JITM class and the register() method call
to the Jetpack::after_plugins_loaded() method. This will ensure that the
JITM class is not used before the plugins have been loaded.

Also remove Automattic\Jetpack\JITM from the autoloader's ignore list.